voice/requirements.txt
openai-whisper==20231117
torch==2.3.0
torchaudio==2.3.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
structlog==24.1.0
